Fix performance comparison and remove entropy number
